refactor: change structure of components
