+++ src/app/app.module.ts
@@ -3,12 +3,13 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import AppComponent from './app.component';
-import HomeComponent from './home/home.component';
 import RoutingModule from './routing/routing.module';
 import MaterialModule from './material/material.module';
-import HeaderComponent from './navigation/header/header.component';
-import SidenavListComponent from './navigation/sidenav-list/sidenav-list.component';
+
+import AppComponent from './app.component';
+import HomeComponent from './pages/home/home.component';
+import HeaderComponent from './components/header/header.component';
+import SidenavListComponent from './components/sidenav-list/sidenav-list.component';
 
 @NgModule({
   declarations: [
