+++ src/app/components/sidenav-list/sidenav-list.component.scss
